+++ packages/common/src/index.ts
@@ -6,4 +6,3 @@ export * as HospitalUtils from "./HospitalUtils";
 export * as BloodTypeUtils from "./BloodTypeUtils";
 export * as DateUtils from "./DateUtils";
 export * as SortingUtils from "./SortingUtils";
-export * as AppointmentUtils from "./AppointmentUtils";
